[REFACTOR] Refactor validation for humidity
